requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
pydantic==2.5.0
streamlit==1.29.0
requests==2.31.0
google-genai==0.3.0
